refactor(various): Integrated the changes from the Improve Profile Picture Upload UX PR into the changes that were associated with the profile page and the profile photo
Also, removed the UpdateProfilePictureForm and its test.
